+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/performance/DefaultPerformanceTracker.java
@@ -16,6 +16,7 @@ package tech.pegasys.teku.validator.coordinator.performance;
 import static com.google.common.base.Preconditions.checkArgument;
 
 import com.google.common.annotations.VisibleForTesting;
+import it.unimi.dsi.fastutil.ints.Int2IntMap;
 import it.unimi.dsi.fastutil.ints.IntArrayList;
 import it.unimi.dsi.fastutil.ints.IntList;
 import it.unimi.dsi.fastutil.ints.IntSet;
@@ -42,7 +43,6 @@ import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.logging.StatusLogger;
 import tech.pegasys.teku.infrastructure.metrics.SettableGauge;
-import tech.pegasys.teku.infrastructure.ssz.collections.SszBitlist;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock;
@@ -51,6 +51,7 @@ import tech.pegasys.teku.spec.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.versions.altair.SyncCommitteeMessage;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
+import tech.pegasys.teku.statetransition.attestation.utils.AttestationBitsAggregator;
 import tech.pegasys.teku.storage.client.CombinedChainDataClient;
 import tech.pegasys.teku.validator.api.ValidatorPerformanceTrackingMode;
 import tech.pegasys.teku.validator.coordinator.ActiveValidatorTracker;
@@ -169,6 +170,9 @@ public class DefaultPerformanceTracker implements PerformanceTracker {
                   validatorPerformanceMetrics.updateBlockPerformanceMetrics(blockPerformance);
                 }
               }
+            })
+        .alwaysRun(
+            () -> {
               producedBlocksByEpoch.headMap(blockProductionEpoch, true).clear();
               blockProductionAttemptsByEpoch.headMap(blockProductionEpoch, true).clear();
             });
@@ -208,8 +212,8 @@ public class DefaultPerformanceTracker implements PerformanceTracker {
                 validatorPerformanceMetrics.updateAttestationPerformanceMetrics(
                     attestationPerformance);
               }
-              producedAttestationsByEpoch.headMap(analyzedEpoch, true).clear();
-            });
+            })
+        .alwaysRun(() -> producedAttestationsByEpoch.headMap(analyzedEpoch, true).clear());
   }
 
   private SafeFuture<BlockPerformance> getBlockPerformanceForEpoch(final UInt64 currentEpoch) {
@@ -300,16 +304,23 @@ public class DefaultPerformanceTracker implements PerformanceTracker {
 
     // Pre-process attestations included on chain to group them by
     // data hash to inclusion slot to aggregation bitlist
-    final Map<Bytes32, NavigableMap<UInt64, SszBitlist>> slotAndBitlistsByAttestationDataHash =
-        new HashMap<>();
+    final Map<Bytes32, NavigableMap<UInt64, AttestationBitsAggregator>>
+        slotAndBitlistsByAttestationDataHash = new HashMap<>();
     for (Map.Entry<UInt64, List<Attestation>> entry : attestationsIncludedOnChain.entrySet()) {
+      final Optional<Int2IntMap> committeesSize =
+          Optional.of(spec.getBeaconCommitteesSize(state, entry.getKey()));
       for (Attestation attestation : entry.getValue()) {
         final Bytes32 attestationDataHash = attestation.getData().hashTreeRoot();
-        final NavigableMap<UInt64, SszBitlist> slotToBitlists =
+        final NavigableMap<UInt64, AttestationBitsAggregator> slotToBitlists =
             slotAndBitlistsByAttestationDataHash.computeIfAbsent(
                 attestationDataHash, __ -> new TreeMap<>());
         slotToBitlists.merge(
-            entry.getKey(), attestation.getAggregationBits(), SszBitlist::nullableOr);
+            entry.getKey(),
+            AttestationBitsAggregator.of(attestation, committeesSize),
+            (firstBitsAggregator, secondBitsAggregator) -> {
+              firstBitsAggregator.or(secondBitsAggregator);
+              return firstBitsAggregator;
+            });
       }
     }
 
@@ -319,10 +330,10 @@ public class DefaultPerformanceTracker implements PerformanceTracker {
       if (!slotAndBitlistsByAttestationDataHash.containsKey(sentAttestationDataHash)) {
         continue;
       }
-      final NavigableMap<UInt64, SszBitlist> slotAndBitlists =
+      final NavigableMap<UInt64, AttestationBitsAggregator> slotAndBitlists =
           slotAndBitlistsByAttestationDataHash.get(sentAttestationDataHash);
       for (UInt64 slot : slotAndBitlists.keySet()) {
-        if (slotAndBitlists.get(slot).isSuperSetOf(sentAttestation.getAggregationBits())) {
+        if (slotAndBitlists.get(slot).isSuperSetOf(sentAttestation)) {
           inclusionDistances.add(slot.minus(sentAttestationSlot).intValue());
           break;
         }
@@ -420,11 +431,11 @@ public class DefaultPerformanceTracker implements PerformanceTracker {
   }
 
   @Override
-  public void saveProducedBlock(final SignedBeaconBlock block) {
-    final UInt64 epoch = spec.computeEpochAtSlot(block.getSlot());
+  public void saveProducedBlock(final SlotAndBlockRoot slotAndBlockRoot) {
+    final UInt64 epoch = spec.computeEpochAtSlot(slotAndBlockRoot.getSlot());
     final Set<SlotAndBlockRoot> blocksInEpoch =
         producedBlocksByEpoch.computeIfAbsent(epoch, __ -> concurrentSet());
-    blocksInEpoch.add(block.getSlotAndBlockRoot());
+    blocksInEpoch.add(slotAndBlockRoot);
   }
 
   @Override