+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/electra/helpers/MiscHelpersElectra.java
@@ -13,9 +13,13 @@
 
 package tech.pegasys.teku.spec.logic.versions.electra.helpers;
 
+import static tech.pegasys.teku.infrastructure.unsigned.UInt64.ZERO;
+import static tech.pegasys.teku.spec.config.SpecConfig.FAR_FUTURE_EPOCH;
+
 import it.unimi.dsi.fastutil.ints.IntList;
 import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.config.SpecConfigDeneb;
 import tech.pegasys.teku.spec.config.SpecConfigElectra;
@@ -69,6 +73,29 @@ public class MiscHelpersElectra extends MiscHelpersDeneb {
     return predicatesElectra.hasCompoundingWithdrawalCredential(validator)
         ? specConfigElectra.getMaxEffectiveBalanceElectra()
         : specConfigElectra.getMinActivationBalance();
+  }
+
+  @Override
+  public Validator getValidatorFromDeposit(
+      final BLSPublicKey pubkey, final Bytes32 withdrawalCredentials, final UInt64 amount) {
+    final Validator validator =
+        new Validator(
+            pubkey,
+            withdrawalCredentials,
+            ZERO,
+            false,
+            FAR_FUTURE_EPOCH,
+            FAR_FUTURE_EPOCH,
+            FAR_FUTURE_EPOCH,
+            FAR_FUTURE_EPOCH);
+
+    final UInt64 maxEffectiveBalance = getMaxEffectiveBalance(validator);
+    final UInt64 validatorEffectiveBalance =
+        amount
+            .minusMinZero(amount.mod(specConfig.getEffectiveBalanceIncrement()))
+            .min(maxEffectiveBalance);
+
+    return validator.withEffectiveBalance(validatorEffectiveBalance);
   }
 
   @Override