+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/P2PConfig.java
@@ -38,7 +38,7 @@ public class P2PConfig {
   public static final int DEFAULT_P2P_TARGET_SUBNET_SUBSCRIBER_COUNT = 2;
   public static final boolean DEFAULT_SUBSCRIBE_ALL_SUBNETS_ENABLED = false;
   public static final boolean DEFAULT_GOSSIP_SCORING_ENABLED = true;
-  public static final boolean DEFAULT_GOSSIP_BLOBS_AFTER_BLOCK_ENABLED = false;
+  public static final boolean DEFAULT_GOSSIP_BLOBS_AFTER_BLOCK_ENABLED = true;
   public static final int DEFAULT_BATCH_VERIFY_MAX_THREADS =
       Math.max(2, Runtime.getRuntime().availableProcessors() / 2);
   public static final int DEFAULT_BATCH_VERIFY_QUEUE_CAPACITY = 15_000;